refactor: relocate text and heading primitives
