+++ src/ui/primitives/Text.tsx
@@ -1,14 +1,38 @@
 import React from 'react';
-import { Text as RNText, StyleProp, TextStyle } from 'react-native';
+import { Text as RNText, TextProps, TextStyle } from 'react-native';
 import { useTheme } from '../ThemeProvider';
 import { typography } from '../tokens';
 
-interface TextProps {
-  children: React.ReactNode;
-  style?: StyleProp<TextStyle>;
+type Variant = keyof typeof typography;
+
+interface Props extends TextProps {
+  variant?: Variant;
+  weight?: TextStyle['fontWeight'];
 }
 
-export default function Text({ children, style }: TextProps) {
-  const { colors } = useTheme();
-  return <RNText style={[{ color: colors.text.primary, ...typography.md }, style]}>{children}</RNText>;
+export default function Text({
+  variant = 'md',
+  weight = '400',
+  style,
+  ...rest
+}: Props) {
+  const { getColor } = useTheme();
+  const { fontSize, lineHeight, letterSpacing } = typography[variant];
+
+  return (
+    <RNText
+      style={[
+        {
+          fontSize,
+          lineHeight,
+          letterSpacing,
+          fontWeight: weight,
+          color: getColor('text.primary'),
+        },
+        style,
+      ]}
+      {...rest}
+    />
+  );
 }
+
