+++ src/ui/primitives/index.ts
@@ -1,15 +1,15 @@
-export { default as Heading } from './Heading';
-export { default as TextField } from './TextField';
-export { default as Card } from './Card';
 export { default as Avatar } from './Avatar';
 export { default as Badge } from './Badge';
+export { default as Button } from './Button';
+export { default as Card } from './Card';
 export { default as Chip } from './Chip';
 export { default as Divider } from './Divider';
-export { default as Portal } from './Portal';
-export { default as Overlay } from './Overlay';
-export { default as Spinner } from './Spinner';
-export { default as Skeleton } from './Skeleton';
+export { default as Heading } from './Heading';
 export { default as Menu } from './Menu';
 export type { MenuItem } from './Menu';
-export { default as Button } from './Button';
+export { default as Overlay } from './Overlay';
+export { default as Portal } from './Portal';
+export { default as Skeleton } from './Skeleton';
+export { default as Spinner } from './Spinner';
 export { default as Text } from './Text';
+export { default as TextField } from './TextField';
